[feat] 모달 기능 subheader 내부에 추가
