+++ src/app/(suwan)/my-study-main/[studyId]/board/_component/NoticeBoardDetail.tsx
@@ -2,10 +2,7 @@
 
 import SubHeader from '@/components/common/SubHeader';
 import Label from '@/components/common/label/Label';
-import { useEffect, useRef, useState } from 'react';
-import { FaCheck } from 'react-icons/fa';
-import KebabModal from '@/components/common/modal/KebabModal';
-import { deleteNotice, postIcon } from '@/utils/my-study-main/fetch';
+import { deleteNotice } from '@/utils/my-study-main/fetch';
 import EmojiSelectBtn from './EmojiSelectBtn';
 import { useRouter } from 'next/navigation';
 import { dateCalculate } from '@/utils/my-study-main/dateCalculate';
@@ -27,29 +24,6 @@ export default function NoticeBoardDetail(props: TProps) {
 
   const boardType = 'notice';
 
-  const [modal, setModal] = useState(false);
-
-  const modalRef = useRef<HTMLDivElement>(null);
-  useEffect(() => {
-    const closeModal = (e: MouseEvent) => {
-      if (
-        modal &&
-        modalRef.current &&
-        !modalRef.current.contains(e.target as Node)
-      ) {
-        // 이벤트가 발생한 노드가 모달 컴포넌트 내부에 존재하지 않는다면 close
-        setModal(false);
-      }
-    };
-
-    // 이벤트 리스너를 document 전체에 붙여줌
-    document.addEventListener('mousedown', closeModal);
-
-    return () => {
-      document.removeEventListener('mousedown', closeModal);
-    };
-  }, [modal]);
-
   const onClickEddit = () => {
     console.log('수정');
     router.push(`./${blackboard._id}/eddit`);
@@ -63,21 +37,14 @@ export default function NoticeBoardDetail(props: TProps) {
   return (
     <>
       {userId === blackboard.writerId._id ? (
-        <SubHeader eddit bgGray onClickMenu={() => setModal(!modal)} />
+        <SubHeader
+          eddit
+          bgGray
+          onClickEddit={onClickEddit}
+          onClickDelete={onClickDelete}
+        />
       ) : (
         <SubHeader />
-      )}
-
-      {modal && (
-        <div
-          ref={modalRef}
-          className="relative inset-0 bg-black bg-opacity-50 z-10"
-        >
-          <KebabModal
-            onClickEddit={onClickEddit}
-            onClickDelete={onClickDelete}
-          />
-        </div>
       )}
       <div className="bg-white p-[2rem] pt-[2.5rem] h-[85vh]">
         <div className="flex justify-between items-center pb-[1.8rem] mb-[1.8rem] border-b border-gray-300">
