+++ src/app/(suwan)/my-study-main/[studyId]/board/_component/TaskDetail.tsx
@@ -4,7 +4,6 @@ import { GoChevronLeft } from 'react-icons/go';
 import Image from 'next/image';
 import EmojiSelectBtn from './EmojiSelectBtn';
 import SubHeader from '@/components/common/SubHeader';
-import { useEffect, useRef, useState } from 'react';
 import { useRouter } from 'next/navigation';
 import { deleteNotice, postIcon } from '@/utils/my-study-main/fetch';
 import KebabModal from '@/components/common/modal/KebabModal';
@@ -22,28 +21,6 @@ export default function TaskDetail(props: TProps) {
   const boardType = 'task';
   const timeAgo = dateCalculate(task.createdAt);
 
-  const [modal, setModal] = useState(false);
-
-  const modalRef = useRef<HTMLDivElement>(null);
-  useEffect(() => {
-    const closeModal = (e: MouseEvent) => {
-      if (
-        modal &&
-        modalRef.current &&
-        !modalRef.current.contains(e.target as Node)
-      ) {
-        // 이벤트가 발생한 노드가 모달 컴포넌트 내부에 존재하지 않는다면 close
-        setModal(false);
-      }
-    };
-    // 이벤트 리스너를 document 전체에 붙여줌
-    document.addEventListener('mousedown', closeModal);
-
-    return () => {
-      document.removeEventListener('mousedown', closeModal);
-    };
-  }, [modal]);
-
   const onClickEddit = () => {
     router.push(`./${task._id}/eddit`);
   };
@@ -56,19 +33,16 @@ export default function TaskDetail(props: TProps) {
   return (
     <>
       {userId === task.writerId._id ? (
-        <SubHeader eddit bgGray onClickMenu={() => setModal(!modal)} />
+        <SubHeader
+          eddit
+          bgGray
+          onClickEddit={onClickEddit}
+          onClickDelete={onClickDelete}
+        />
       ) : (
         <SubHeader />
       )}
-      {modal && (
-        <div ref={modalRef} className="relative inset-0 z-10">
-          <KebabModal
-            eddit
-            onClickEddit={onClickEddit}
-            onClickDelete={onClickDelete}
-          />
-        </div>
-      )}
+
       <div className="bg-white px-[1.6rem] flex flex-col gap-y-[2rem] py-[3rem]">
         <div className="flex flex-row items-center">
           <Image
